refactor: extract MockIcons to shared fixture in conftest.py
- Added mock_ui_icons, mock_ui_menu_option, and mock_ui_module_full fixtures to conftest.py
- Updated test_cli_input_handlers.py and test_cli_utils.py to use shared fixtures
- Removed module-level sys.modules['ui'] mutations in favor of autouse fixtures
- Removed duplicated MockIcons, MockMenuOption, and helper function definitions
- All 497 tests pass with 3 skipped (subprocess tests require claude_agent_sdk)

This addresses CodeRabbit feedback about code duplication and sys.modules
pollution across test files. The shared fixture approach improves maintainability
and ensures proper cleanup between test runs.
